refactor: Team page transferred  to About page fix: Fixed error with category page displaying
