+++ src/index.tsx
@@ -23,7 +23,6 @@ import ProductsPage from "@/pages/products/ProductsPage";
 import SearchResultsPage from "@/pages/search/SearchResultsPage";
 import CategoryProductsPage from "@/pages/category/CategoryProductsPage";
 import CartPage from "@/pages/cart/CartPage";
-import TeamPage from "@/pages/team/TeamPage";
 
 const root = document.getElementById("root");
 
@@ -67,14 +66,6 @@ const router = createBrowserRouter([
         element: (
           <Suspense fallback={"Loading..."}>
             <AboutPage />
-          </Suspense>
-        ),
-      },
-      {
-        path: "/team",
-        element: (
-          <Suspense fallback={"Loading..."}>
-            <TeamPage />
           </Suspense>
         ),
       },
